pyproject.toml
[project]
name = "mastodondownloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mastodon-py>=1.8.1",
    "peewee>=3.17.8",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
